Fix filename issue for files that have moved
For each line in a file, git blame reports the name of the file as it
was when the line was committed. If the file has moved, this will be
different from the actual file name. We now set all filenames to be the
current name of the file which lets us reconstruct the file from just
the JSON.